refactor: Split server explosion implementation
Organize explosion orchestration around Vanilla's ray, entity, exposure, block-effect, and damage-source phases without changing runtime behavior.
